+++ app/lib/auth.ts
@@ -1,5 +1,6 @@
 import * as SecureStore from "expo-secure-store";
-
+import * as Linking from "expo-linking";
+import { fetchAPI } from "./fetch";
 export interface TokenCache {
   getToken: (key: string) => Promise<string | undefined | null>;
   saveToken: (key: string, token: string) => Promise<void>;
@@ -29,4 +30,48 @@ export const tokenCache = {
       return;
     }
   },
+};
+
+export const googleOAuth = async (startOAuthFlow: any) => {
+  try {
+    const { createdSessionId, signUp, setActive } = await startOAuthFlow({
+      redirectUrl: Linking.createURL("/(root)/(tabs)/home", {
+        scheme: "myapp",
+      }),
+    });
+
+    if (createdSessionId) {
+      if (setActive) {
+        setActive!({ session: createdSessionId });
+        if (signUp.createdUserId) {
+          await fetchAPI("/(api)/user", {
+            method: "POST",
+            body: JSON.stringify({
+              name: `${signUp.firstName} ${signUp.lastName}`,
+              email: signUp.emailAddress,
+              clerkId: signUp.createdUserId,
+            }),
+          });
+        }
+
+        return {
+          success: true,
+          code: "success",
+          message: "You have successfully authenticated",
+        };
+      }
+    }
+    return {
+      success: false,
+      code: "success",
+      message: "An error occurred",
+    };
+  } catch (error: any) {
+    console.log(error);
+    return {
+      success: false,
+      code: error.code,
+      message: error?.errors[0]?.longMessage || "An error occurred",
+    };
+  }
 };