+++ components/OAuth.tsx
@@ -1,9 +1,28 @@
-import { Image, Text, View } from "react-native";
+import { Alert, Image, Text, View } from "react-native";
 import CustomButton from "./CustomButton";
 import { icons } from "@/constants";
+import { useOAuth } from "@clerk/clerk-expo";
+import { useCallback } from "react";
+import { googleOAuth } from "@/app/lib/auth";
+import { router } from "expo-router";
 
 const OAuth = () => {
-  const handleGoogleSignIn = () => {};
+  const { startOAuthFlow } = useOAuth({ strategy: "oauth_google" });
+
+  const handleGoogleSignIn = useCallback(async () => {
+    try {
+      const result = await googleOAuth(startOAuthFlow);
+
+      if (result.code === "session_exists") {
+        Alert.alert("Success", "Session Exists, Redirecting to home page");
+        router.push("/(root)/(tabs)/home");
+      }
+
+      Alert.alert(result.success ? "Success" : "Error", result.message);
+    } catch (err) {
+      console.error("OAuth error", err);
+    }
+  }, []);
 
   return (
     <View>